requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.0
llama-cpp-python==0.2.72
jinja2==3.1.3
